Time: 4 ms (38.72%) | Memory: 8.4 MB (10.33%) - LeetSync

+++ 1833-find-the-highest-altitude/find-the-highest-altitude.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int largestAltitude(vector<int>& gain) {
-        short int a=0,max=0;
+        int a=0,max=0;
         for(auto i:gain)
         {
             a+=i;
